Moved constant declarations to constants.dart
