+++ Flutter/lib/constants.dart
@@ -1,2 +1,235 @@
-double numBtnHeigth = 100.0;
-double numBtnWidth = 100.0;
+// ignore_for_file: constant_identifier_names
+
+import 'package:flutter/material.dart';
+
+const PRIMARY_COLOR = 0xFF4A0000;
+const DEFAULT_BTN_COLOR = 0xFF4A0000;
+const BACKGROUND_COLOR = 0xFF181818;
+
+double numBtnWidth = 26.04;
+double numBtnHeigth = 8.57;
+double sideBtnWidth = 17.71;
+double sideBtnHeigth = 8.58;
+
+enum ChosenPlayerEnum { player1, player2, undefined }
+
+int serverPort = 11520;
+const List<Color> rankingColors = [
+  Color.fromRGBO(255, 215, 0, 1.0),
+  Color.fromRGBO(192, 192, 192, 1.0),
+  Color.fromRGBO(128, 128, 128, 1.0)
+];
+
+var sideBtnStyle = ElevatedButton.styleFrom(
+  shape: const RoundedRectangleBorder(
+    borderRadius: BorderRadius.all(Radius.zero),
+  ),
+  elevation: 0,
+  shadowColor: Colors.transparent,
+  primary: Colors.grey[850],
+);
+var numBtnStyle = ElevatedButton.styleFrom(
+  side: const BorderSide(
+      color: Colors.black, width: 2.0, style: BorderStyle.solid),
+  shape: const RoundedRectangleBorder(
+    borderRadius: BorderRadius.all(Radius.zero),
+  ),
+  primary: Color(DEFAULT_BTN_COLOR),
+);
+var cBtnStyle = ElevatedButton.styleFrom(
+  side: const BorderSide(
+      color: Colors.black, width: 2.0, style: BorderStyle.solid),
+  shape: const RoundedRectangleBorder(
+    borderRadius: BorderRadius.all(Radius.zero),
+  ),
+  primary: const Color(0xFFC40000),
+);
+var bustBtnSytle = ElevatedButton.styleFrom(
+  side: const BorderSide(
+      color: Colors.black, width: 2.0, style: BorderStyle.solid),
+  shape: const RoundedRectangleBorder(
+    borderRadius: BorderRadius.all(Radius.zero),
+  ),
+  primary: const Color(0xFF9E5905),
+);
+var okBtnStyle = ElevatedButton.styleFrom(
+  side: const BorderSide(
+      color: Colors.black, width: 2.0, style: BorderStyle.solid),
+  shape: const RoundedRectangleBorder(
+    borderRadius: BorderRadius.all(Radius.zero),
+  ),
+  primary: const Color(0xFF014D05),
+);
+var numPadFontSize = 30;
+
+List<String> possibleOuts = [
+  'T20 T20 BULL',
+  '',
+  '',
+  'T20 T19 BULL',
+  '',
+  '',
+  'T20 T18 BULL',
+  '',
+  '',
+  'T20 T17 BULL',
+  'T20 T20 D20',
+  '',
+  'T20 T20 D19',
+  'T20 T19 D20',
+  'T20 T20 D18',
+  'T20 T19 D19',
+  'T20 T18 D20',
+  'T20 T19 D18',
+  'T20 T20 D16',
+  'T20 T17 D20',
+  'T20 T18 D18',
+  'T20 T19 D16',
+  'T20 T16 D20',
+  'T20 T17 D18',
+  'T20 T18 D16',
+  'T20 T15 D20',
+  'T20 T20 D12',
+  'T20 T17 D16',
+  'T20 T14 D20',
+  'T20 T19 D12',
+  'T20 T16 D16',
+  'T19 T14 D20',
+  'T20 T18 D12',
+  'T19 T16 D16',
+  'T20 T20 D8',
+  'T20 T17 D12',
+  'T20 T14 D16',
+  'T20 T19 D8',
+  'BULL T14 D20',
+  'T20 T13 D16',
+  'T20 T20 D5',
+  'T19 T20 D6',
+  'T18 T14 D16',
+  'T20 T17 D8',
+  'T19 T19 D6',
+  '25 T20 D20',
+  'T20 T14 D11',
+  'T19 T16 D9',
+  'T18 T18 D7',
+  'T20 T11 D14',
+  'T20 20 D20',
+  'T19 T12 D13',
+  'T20 18 D20',
+  'T20 17 D20',
+  'T20 16 D20',
+  'T20 15 D20',
+  'T20 14 D20',
+  'T20 13 D20',
+  'T20 12 D20',
+  'T20 11 D20',
+  'T20 10 D20',
+  'T19 12 D20',
+  'T20 16 D16',
+  'T19 10 D20',
+  'T20 10 D18',
+  'T20 13 D16',
+  'T20 12 D16',
+  'T19 10 D18',
+  'T20 10 D16',
+  'T17 10 D20',
+  'T20 D20',
+  'T19 10 D16',
+  'T20 D19',
+  'T19 D20',
+  'T20 D18',
+  'T19 D19',
+  'T18 D20',
+  'T19 D18',
+  'T20 D16',
+  'T17 D20',
+  'T20 D15',
+  'T19 D16',
+  'T20 D14',
+  'T17 D18',
+  'T18 D16',
+  'T15 D20',
+  'T16 D18',
+  'T17 D16',
+  'T14 D20',
+  'T15 D18',
+  'T20 D10',
+  'T13 D20',
+  'T18 D12',
+  'T15 D16',
+  'T20 D8',
+  'T13 D18',
+  'T14 D16',
+  'T19 D8',
+  'T16 D12',
+  'T13 D16',
+  'T18 D8',
+  'T19 D6',
+  'T20 D4',
+  'T17 D8',
+  'T10 D18',
+  'T19 D4',
+  'T16 D8',
+  'T13 D12',
+  'T10 D16',
+  'T15 D8',
+  '20 D20',
+  '19 D20',
+  '18 D20',
+  '17 D20',
+  '16 D20',
+  '15 D20',
+  '14 D20',
+  '13 D20',
+  '12 D20',
+  '19 D16',
+  '10 D20',
+  '17 D16',
+  '16 D16',
+  '15 D16',
+  '6 D20',
+  '13 D16',
+  '12 D16',
+  '3 D20',
+  '10 D16',
+  '9 D16',
+  'D20',
+  '7 D16',
+  'D19',
+  '5 D16',
+  'D18',
+  '3 D16',
+  'D17',
+  '1 D16',
+  'D16',
+  '15 D8',
+  'D15',
+  '13 D8',
+  'D14',
+  '19 D4',
+  'D13',
+  '9 D8',
+  'D12',
+  '7 D8',
+  'D11',
+  '5 D8',
+  'D10',
+  '3 D8',
+  'D9',
+  '9 D4',
+  'D8',
+  '7 D4',
+  'D7',
+  '5 D4',
+  'D6',
+  '3 D4',
+  'D5',
+  '1 D4',
+  'D4',
+  '3 D2',
+  'D3',
+  '1 D2',
+  'D2',
+  '1 D1',
+  'D1'
+];
